+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 7.1.0 (IN PROGRESS)
 
 * Add "Donor (Deprecated)" label for PO line search by Donor filed. Refs UIOR-1348.
+* Provide system only POL changed fields for POL version history. Refs UIOR-1362.
 
 ## [7.0.1](https://github.com/folio-org/ui-orders/tree/v7.0.1) (2024-11-14)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v7.0.0...v7.0.1)